Simplify read_random_bytes.

+++ kernel.h
@@ -77,6 +77,6 @@ int kernel_addresses(char *ifname, int ifindex, int ll,
                      struct kernel_route *routes, int maxroutes);
 int if_eui64(char *ifname, int ifindex, unsigned char *eui);
 int gettime(struct timeval *tv);
-int read_random_bytes(void *buf, size_t len);
+ssize_t read_random_bytes(void *buf, size_t len);
 int kernel_older_than(const char *sysname, int version, int sub_version);
 int kernel_has_ipv6_subtrees(void);
